Extract ini file and directory functions into PhpBinaryPath
